+++ Classes/Model/AdditionalProperties/Drug.php
@@ -27,7 +27,6 @@ final class Drug implements AdditionalPropertiesInterface
             'colorSwatch',
             'countryOfAssembly',
             'countryOfLastProcessing',
-            'depth',
             'funding',
             'gtin',
             'hasAdultConsideration',
@@ -35,7 +34,6 @@ final class Drug implements AdditionalPropertiesInterface
             'hasGS1DigitalLink',
             'hasMeasurement',
             'hasMerchantReturnPolicy',
-            'height',
             'inProductGroupWithID',
             'includedInHealthInsurancePlan',
             'mobileUrl',
@@ -45,8 +43,6 @@ final class Drug implements AdditionalPropertiesInterface
             'positiveNotes',
             'rxcui',
             'size',
-            'weight',
-            'width',
         ];
     }
 }