+++ Classes/Model/AdditionalProperties/MusicVideoObject.php
@@ -37,7 +37,6 @@ final class MusicVideoObject implements AdditionalPropertiesInterface
             'editEIDR',
             'educationalLevel',
             'funding',
-            'height',
             'ineligibleRegion',
             'interpretedAsClaim',
             'maintainer',
@@ -51,7 +50,6 @@ final class MusicVideoObject implements AdditionalPropertiesInterface
             'size',
             'teaches',
             'usageInfo',
-            'width',
         ];
     }
 }